+++ code/server/SmartHome/src/cn/fuego/smart/home/domain/FireSensor.java
@@ -17,25 +17,26 @@ package cn.fuego.smart.home.domain;
  */
 public class FireSensor
 {
-	private int id;
-	private int concentratorID;  //集中器ID
+	private long id;
+	private long concentratorID;  //集中器ID
 
 	private int machineID; //机号
 	private int loopID;    //回路号
 	private int codeID;    //编号
-	public int getId()
+
+	public long getId()
 	{
 		return id;
 	}
-	public void setId(int id)
+	public void setId(long id)
 	{
 		this.id = id;
 	}
-	public int getConcentratorID()
+	public long getConcentratorID()
 	{
 		return concentratorID;
 	}
-	public void setConcentratorID(int concentratorID)
+	public void setConcentratorID(long concentratorID)
 	{
 		this.concentratorID = concentratorID;
 	}
